Move strategies config parts into subdir
This is @joelva suggestion to not mess strategies and their config
elements in the same dir.
