+++ dexbot/strategies/config_parts/base_config.py
